+++ NEWS.md
@@ -4,7 +4,7 @@
 
 # microinverterdata 0.1.2
 
-* Add mocked http data for APSystems and Fronius #2
+* Add mocked http data for APSystems and Fronius #4
 * Add support for 'Fronius' inverters 
 
 # microinverterdata 0.1.1